typersonal/ai_requirements.txt
accelerate==0.23.0
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
attrs==25.3.0
booleanOperations==0.9.0
certifi==2025.1.31
cffsubr==0.3.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
compreffor==0.5.6
contourpy==1.3.0
cycler==0.12.1
diffusers==0.32.2
exceptiongroup==1.2.2
fastapi==0.99.1
ffmpy==0.5.0
filelock==3.18.0
fontmake==3.10.0
fontMath==0.9.4
fonttools==4.56.0
fs==2.4.16
fsspec==2025.3.0
glyphsLib==6.10.1
gradio==3.50.2
gradio_client==0.6.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
info-nce-pytorch==0.1.4
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
kornia==0.8.0
kornia_rs==0.1.8
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.31.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.11.0.86
openstep_plist==0.5.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==10.4.0
psutil==7.0.0
pyclipper==1.3.0.post6
pydantic==1.10.21
pydantic_core==2.27.2
pydub==0.25.1
pygame==2.6.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
safetensors==0.5.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
svgwrite==1.4.3
sympy==1.13.1
tokenizers==0.13.3
tomlkit==0.12.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.33.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
ufo2ft==3.4.2
ufoLib2==0.17.1
unicodedata2==16.0.0
urllib3==2.3.0
uvicorn==0.34.0
websockets==11.0.3
zipp==3.21.0
